backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Scraping
scrapling==0.4.2
browserforge==1.2.4
playwright==1.58.0
curl-cffi==0.14.0

# AI/ML
torch>=2.5.1
transformers==5.3.0
accelerate==1.13.0
json-repair==0.25.0

# Utilities
pydantic==2.12.5
python-dotenv==1.0.0
requests==2.31.0
modal==1.3.5

# Part 2 market data pipeline
pandas>=2.2.3
numpy>=2.1.0
yfinance==0.2.54
